fixed get events revenue

+++ event-ticketing-system/src/main/java/com/example/event_ticketing_system/dto/RevenueDTO.java
@@ -1,5 +1,7 @@
 package com.example.event_ticketing_system.dto;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -9,5 +11,5 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class RevenueDTO {
     private String eventTitle;
-    private Integer totalConfirmedRevenue;
+    private BigDecimal  totalConfirmedRevenue;
 }
